==> 2_data_analysis/5_central_user_network_analyze.ipynb <==
import pandas as pd

df = pd.read_feather("./data/10_analy_top_nodes_acc_post.feather")

df

df_select = df[
    [
        "account_id",
        "language_type",
        "sentiment_class",
        "topic_sent_num_posts",
        "account_url",
    ]
]

df_select

acct_num_posts = (
    df_select.groupby(["account_id"])["topic_sent_num_posts"]
    .sum()
    .reset_index(name="acct_num_posts")
)

acct_num_posts

acct_sent_num_posts = (
    df_select.groupby(["account_id", "sentiment_class"])["topic_sent_num_posts"]
    .sum()
    .reset_index(name="acct_sent_num_posts")
)

acct_sent_pct_posts = pd.merge(acct_num_posts, acct_sent_num_posts, on="account_id")
acct_sent_pct_posts["account_sent_pct_posts"] = (
    acct_sent_pct_posts["acct_sent_num_posts"] / acct_sent_pct_posts["acct_num_posts"]
)

acct_sent_pct_posts

acct_sent_info = acct_sent_pct_posts.pivot(
    index="account_id",
    columns="sentiment_class",
    values="account_sent_pct_posts",
)
acct_sent_info = acct_sent_info.fillna(0)
acct_sent_info = acct_sent_info[["Negative", "Neutral", "Positive"]].rename(
    columns={
        "Negative": "Negative posts",
        "Neutral": "Neutral posts",
        "Positive": "Positive posts",
    }
)
acct_sent_info = acct_sent_info.reset_index()
acct_sent_info.columns.name = None

acct_sent_info

acct_key_info = df.drop(
    [
        "sum_reblogs_count",
        "sum_favourites_count",
        "sum_replies_count",
        "topic_name",
        "sentiment_class",
        "topic_sent_num_posts",
    ],
    axis=1,
)
acct_key_info["instance"] = acct_key_info["instance"].apply(lambda x: x[0]).astype(str)
acct_key_info = acct_key_info.drop_duplicates().reset_index(drop=True)

acct_key_info

final_acct_info = pd.merge(acct_key_info, acct_sent_info, on="account_id")

final_acct_info

final_acct_info.to_excel(
    "./data/10_analy_top_nodes_acc_post_sentiments.xlsx", index=False
)
